Add prompt box and use conditional statements.
I have utilized the prompt method in the function getHumanChoice with a prompt message to prompt the user to input their choice. I have further declared a variable where the user's input will be stored.  I have used conditional statements to ensure that the function will return any of the valid choices that the user inputs.